The stations have to be nested inside the network XML tag. Therefore, the network_code element of the station is no longer needed.
Removed the camel case naming of the assignedSensorUnit.

Changed the default value of tf_poles and tf_zeros to None to prevent weird 
behaviour of the tf_poles and tf_zeros list. With the empty list as the default, the 
list was initiated with the list of the prior SensorParameter instance.

Added a logger to the SensorParameter class.

Added a method to export an inventory to XML file.