Removed the create ppsd images collection node.
I think that this is an old collection node which I used before
the looper node was available. The computation of the PPSD data
and images is currently handled by the compute_ppsd looper child.
For the future, the looper child should only create the PPSD data
and the create_ppsd_images collection node should be used to create
the images using this data.